# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de gestión escolar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  rut: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  alumno: 'alumno',
  apoderado: 'apoderado',
  profesor: 'profesor',
  inspector: 'inspector',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  rut: string;
  nombre: string;
  apellido: string;
  email: string;
  role: UserRole;
  /** @nullable */
  telefono?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  alumno: 'alumno',
  apoderado: 'apoderado',
  profesor: 'profesor',
  inspector: 'inspector',
  admin: 'admin',
} as const;

export interface UserInput {
  rut: string;
  nombre: string;
  apellido: string;
  email: string;
  role: UserInputRole;
  password: string;
  telefono?: string;
  /** @nullable */
  apoderadoId?: number | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  alumno: 'alumno',
  apoderado: 'apoderado',
  profesor: 'profesor',
  inspector: 'inspector',
  admin: 'admin',
} as const;

export interface UserUpdate {
  nombre?: string;
  apellido?: string;
  email?: string;
  telefono?: string;
  role?: UserUpdateRole;
}

export interface Course {
  id: number;
  nombre: string;
  /** @nullable */
  descripcion?: string | null;
  profesorId: number;
  profesorNombre: string;
}

export interface CourseInput {
  nombre: string;
  descripcion?: string;
  profesorId: number;
}

export interface ScheduleEntry {
  id: number;
  courseId: number;
  courseName: string;
  teacherName: string;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  classroom: string;
}

export interface ScheduleInput {
  courseId: number;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  classroom: string;
}

export interface Grade {
  id: number;
  studentId: number;
  studentName: string;
  studentRut: string;
  courseId: number;
  courseName: string;
  evaluationName: string;
  value: number;
  createdAt?: string;
}

export interface GradeInput {
  studentId: number;
  courseId: number;
  evaluationName: string;
  value: number;
}

export interface GradeUpdate {
  evaluationName?: string;
  value?: number;
}

export interface GradeAverage {
  studentId: number;
  courseId: number;
  average: number;
  totalGrades: number;
}

export interface AttendanceRecord {
  id: number;
  studentId: number;
  studentName: string;
  courseId: number;
  courseName: string;
  date: string;
  present: boolean;
}

export interface AttendanceInput {
  studentId: number;
  courseId: number;
  date: string;
  present: boolean;
}

export interface AttendanceUpdate {
  present?: boolean;
}

export type AnnotationType = typeof AnnotationType[keyof typeof AnnotationType];


export const AnnotationType = {
  positiva: 'positiva',
  negativa: 'negativa',
} as const;

export interface Annotation {
  id: number;
  studentId: number;
  studentName: string;
  inspectorId: number;
  inspectorName: string;
  type: AnnotationType;
  description: string;
  date: string;
}

export type AnnotationInputType = typeof AnnotationInputType[keyof typeof AnnotationInputType];


export const AnnotationInputType = {
  positiva: 'positiva',
  negativa: 'negativa',
} as const;

export interface AnnotationInput {
  studentId: number;
  inspectorId: number;
  type: AnnotationInputType;
  description: string;
}

export type AnnotationUpdateType = typeof AnnotationUpdateType[keyof typeof AnnotationUpdateType];


export const AnnotationUpdateType = {
  positiva: 'positiva',
  negativa: 'negativa',
} as const;

export interface AnnotationUpdate {
  type?: AnnotationUpdateType;
  description?: string;
}

export type MeetingStatus = typeof MeetingStatus[keyof typeof MeetingStatus];


export const MeetingStatus = {
  pendiente: 'pendiente',
  confirmada: 'confirmada',
  cancelada: 'cancelada',
} as const;

export interface Meeting {
  id: number;
  requestedById: number;
  requestedByName: string;
  withUserId: number;
  withUserName: string;
  title: string;
  /** @nullable */
  description?: string | null;
  scheduledAt: string;
  status: MeetingStatus;
}

export interface MeetingInput {
  requestedById: number;
  withUserId: number;
  title: string;
  description?: string;
  scheduledAt: string;
}

export type MeetingUpdateStatus = typeof MeetingUpdateStatus[keyof typeof MeetingUpdateStatus];


export const MeetingUpdateStatus = {
  pendiente: 'pendiente',
  confirmada: 'confirmada',
  cancelada: 'cancelada',
} as const;

export interface MeetingUpdate {
  title?: string;
  description?: string;
  scheduledAt?: string;
  status?: MeetingUpdateStatus;
}

export type AnnouncementType = typeof AnnouncementType[keyof typeof AnnouncementType];


export const AnnouncementType = {
  anuncio: 'anuncio',
  evento: 'evento',
  festividad: 'festividad',
  alianza: 'alianza',
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  type: AnnouncementType;
  createdAt: string;
  authorName: string;
  /** @nullable */
  eventDate?: string | null;
}

export type AnnouncementInputType = typeof AnnouncementInputType[keyof typeof AnnouncementInputType];


export const AnnouncementInputType = {
  anuncio: 'anuncio',
  evento: 'evento',
  festividad: 'festividad',
  alianza: 'alianza',
} as const;

export interface AnnouncementInput {
  title: string;
  content: string;
  type: AnnouncementInputType;
  authorId: number;
  eventDate?: string;
}

export type AnnouncementUpdateType = typeof AnnouncementUpdateType[keyof typeof AnnouncementUpdateType];


export const AnnouncementUpdateType = {
  anuncio: 'anuncio',
  evento: 'evento',
  festividad: 'festividad',
  alianza: 'alianza',
} as const;

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  type?: AnnouncementUpdateType;
  eventDate?: string;
}

export type StudentDashboardGradesSummaryItem = {
  courseId: number;
  courseName: string;
  average: number;
};

export interface StudentDashboard {
  totalCourses: number;
  upcomingMeetings: number;
  recentAnnouncements: number;
  gradesSummary: StudentDashboardGradesSummaryItem[];
}

export interface TeacherDashboard {
  totalStudents: number;
  totalCourses: number;
  pendingMeetings: number;
  recentGrades: number;
}

export interface AdminDashboard {
  totalAlumnos: number;
  totalApoderados: number;
  totalProfesores: number;
  totalInspectores: number;
  totalMeetings: number;
  totalAnnouncements: number;
  totalCourses: number;
}

export type ListUsersParams = {
role?: ListUsersRole;
};

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  alumno: 'alumno',
  apoderado: 'apoderado',
  profesor: 'profesor',
  inspector: 'inspector',
  admin: 'admin',
} as const;

export type ListScheduleParams = {
studentId?: number;
teacherId?: number;
courseId?: number;
};

export type ListGradesParams = {
studentId?: number;
courseId?: number;
rut?: string;
};

export type ListAttendanceParams = {
courseId?: number;
studentId?: number;
};

export type ListAnnotationsParams = {
studentId?: number;
type?: ListAnnotationsType;
};

export type ListAnnotationsType = typeof ListAnnotationsType[keyof typeof ListAnnotationsType];


export const ListAnnotationsType = {
  positiva: 'positiva',
  negativa: 'negativa',
} as const;

export type ListMeetingsParams = {
requestedById?: number;
withUserId?: number;
};