# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema de gestión escolar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  # SCHEDULE (Horario)
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ANNOTATIONS (Anotaciones)
  /annotations:
    get:
      operationId: listAnnotations
      tags: [annotations]
      summary: List annotations
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [positiva, negativa]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Annotation"
    post:
      operationId: createAnnotation
      tags: [annotations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
  /annotations/{id}:
    patch:
      operationId: updateAnnotation
      tags: [annotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
    delete:
      operationId: deleteAnnotation
      tags: [annotations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # MEETINGS (Reuniones)
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements and events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DASHBOARD SUMMARIES
  /dashboard/student/{studentId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard summary
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
  /dashboard/teacher/{teacherId}:
    get:
      operationId: getTeacherDashboard
      tags: [dashboard]
      summary: Get teacher dashboard summary
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboard"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LoginInput:
      type: object
      required: [rut, password]
      properties:
        rut:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, rut, nombre, apellido, role, email]
      properties:
        id:
          type: integer
        rut:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [alumno, apoderado, profesor, inspector, admin]
        telefono:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Course:
      type: object
      required: [id, nombre, profesorId, profesorNombre]
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        profesorId:
          type: integer
        profesorNombre:
          type: string
    CourseInput:
      type: object
      required: [nombre, profesorId]
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        profesorId:
          type: integer
...
    AttendanceRecord:
      type: object
      required: [id, studentId, studentName, courseId, courseName, date, present]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        date:
          type: string
        present:
          type: boolean
    AttendanceInput:
      type: object
      required: [studentId, courseId, date, present]
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        date:
          type: string
        present:
          type: boolean
    AttendanceUpdate:
      type: object
      properties:
        present:
          type: boolean
    Annotation:
      type: object
      required: [id, studentId, studentName, inspectorId, inspectorName, type, description, date]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        inspectorId:
          type: integer
        inspectorName:
          type: string
        type:
          type: string
          enum: [positiva, negativa]
        description:
          type: string
        date:
          type: string
    AnnotationInput:
      type: object
      required: [studentId, inspectorId, type, description]
      properties:
        studentId:
          type: integer
        inspectorId:
          type: integer
        type:
          type: string
          enum: [positiva, negativa]
        description:
          type: string
    AnnotationUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [positiva, negativa]
        description:
          type: string
...
    Announcement:
      type: object
      required: [id, title, content, type, createdAt, authorName]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [anuncio, evento, festividad, alianza]
        createdAt:
          type: string
        authorName:
          type: string
        eventDate:
          type: ["string", "null"]
    AnnouncementInput:
      type: object
      required: [title, content, type, authorId]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [anuncio, evento, festividad, alianza]
        authorId:
          type: integer
        eventDate:
          type: string
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [anuncio, evento, festividad, alianza]
        eventDate:
          type: string
    StudentDashboard:
      type: object
      required: [totalCourses, upcomingMeetings, recentAnnouncements, gradesSummary]
      properties:
        totalCourses:
          type: integer
        upcomingMeetings:
          type: integer
        recentAnnouncements:
          type: integer
        gradesSummary:
          type: array
          items:
            type: object
            required: [courseId, courseName, average]
            properties:
              courseId:
                type: integer
              courseName:
                type: string
              average:
                type: number
    TeacherDashboard:
      type: object
      required: [totalStudents, totalCourses, pendingMeetings, recentGrades]
      properties:
        totalStudents:
          type: integer
        totalCourses:
          type: integer
        pendingMeetings:
          type: integer
        recentGrades:
          type: integer
...